- decomposited router index. - replaced routes to routes.ts. - updated version.
