+++ package.json
@@ -2,7 +2,7 @@
   "name": "wavelovers",
   "description": "Wavelovers. Use your device vibration correctly. Make a massager out of a gamepad.",
   "keywords": [ "wavelovers", "gamepad-vibrator", "gamepad-test-tool", "gamepad-vibration-test-tool" ],
-  "version": "1.0.10",
+  "version": "1.0.11",
   "license": "GNU GPL v3",
   "homepage": "https://wavelovers.ru/",
   "author": {
